moved producers code under same path
